implementção das funções realizadas anteriormente para o ficheiro emprestimos.py
